Separate types into more specific files.
