+++ src/features/photos/PhotoViewer/types.ts
@@ -1,0 +1,7 @@
+import { Rover, RoverPhotoStateObj } from "../../../types/common";
+
+export interface RoverPhotoState {
+  photos: RoverPhotoStateObj;
+  selectedRover: Rover;
+  page: number;
+}
